embedding_multimodal/requirements.txt
# Bibliotecas para o Pipeline de Embedding Multimodal

# ── Core & API ──
python-dotenv>=1.0.1
openai>=2.16.0

# ── Processamento Multimodal (PDF/Imagens/Tabelas) ──
pymupdf>=1.27.1
Pillow>=12.0.0
pandas>=2.2.0
tabulate>=0.9.0

# ── Vetores e Armazenamento ──
faiss-cpu>=1.9.0
chromadb>=1.5.0
numpy>=2.3.5

# ── Refinamento ──
sentence-transformers>=5.2.3

# ── Orquestração de RAG & LLM ──
langchain>=1.2.9
langchain-openai>=1.1.7
langgraph>=1.0.7
pydantic>=2.9.0
